Move IrUtils to common package
