+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/serialization/SerializerVisitor.kt
@@ -16,25 +16,15 @@
 
 package org.jetbrains.kotlin.backend.konan.ir
 
+import org.jetbrains.kotlin.backend.common.ir.ir2stringWhole
 import org.jetbrains.kotlin.backend.konan.Context
 import org.jetbrains.kotlin.descriptors.FunctionDescriptor
-import org.jetbrains.kotlin.serialization.deserialization.descriptors.*
-import org.jetbrains.kotlin.descriptors.*
-import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.declarations.*
-import org.jetbrains.kotlin.ir.descriptors.IrBuiltIns
-import org.jetbrains.kotlin.ir.visitors.IrElementVisitorVoid
-import org.jetbrains.kotlin.ir.visitors.acceptChildrenVoid
-import org.jetbrains.kotlin.ir.visitors.acceptVoid
 import org.jetbrains.kotlin.backend.konan.descriptors.EmptyDescriptorVisitorVoid
 import org.jetbrains.kotlin.backend.konan.descriptors.*
 import org.jetbrains.kotlin.backend.konan.serialization.*
-import org.jetbrains.kotlin.backend.konan.llvm.base64Encode
-import org.jetbrains.kotlin.backend.konan.llvm.base64Decode
 import org.jetbrains.kotlin.backend.konan.PhaseManager
 import org.jetbrains.kotlin.backend.konan.KonanPhase
-import org.jetbrains.kotlin.serialization.MutableTypeTable
-import org.jetbrains.kotlin.serialization.KonanLinkData
 
 internal class DeserializerDriver(val context: Context) {
 
@@ -48,7 +38,7 @@ internal class DeserializerDriver(val context: Context) {
             try {
                 deserializedIr = IrDeserializer(context, descriptor).decodeDeclaration()
                 context.log{"${deserializedIr!!.descriptor}"}
-                context.log{ir2stringWhole(deserializedIr!!)}
+                context.log{ ir2stringWhole(deserializedIr!!) }
                 context.log{"IR deserialization SUCCESS:\t$descriptor"}
             } catch(e: Throwable) {
                 context.log{"IR deserialization FAILURE:\t$descriptor"}
